Movie MyApp to seperate class
