+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:mixeditems/models/listitem.dart';
 import 'package:mixeditems/repository/repo.dart';
+import 'package:mixeditems/screens/home.dart';
 
 void main() {
   List<ListItem> listItems = Repo().getListItems();
@@ -16,36 +17,6 @@ void main() {
 }
 
 
-class MyApp extends StatelessWidget {
-  final List<ListItem> items;
-
-  const MyApp({Key? key, required this.items}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    const title = 'Mixed List';
-
-    return MaterialApp(
-      title: title,
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text(title),
-        ),
-        body: ListView.builder(
-          itemCount: items.length,
-          itemBuilder: (context, index) {
-            final item = items[index];
-
-            return ListTile(
-              visualDensity: VisualDensity(horizontal: 0, vertical: -4),
-              title:item.buildBody(context),
-            );
-          },
-        ),
-      ),
-    );
-  }
-}
